perf: isolate friends galaxy engine and transfer scene buffers
